refactor: integrate the mistral models testing script to test_chat
